requirements.txt
numpy~=1.26.4
torch~=2.2.0
transformers~=4.38.1
tqdm~=4.66.2
scikit-learn~=1.4.0
numba~=0.59.0
datasets~=2.17.0
accelerate~=0.27.2
setuptools~=68.2.0
pandas~=2.2.0
safetensors~=0.4.2
threadpoolctl~=3.2.0
pyyaml~=6.0.1
